move "openshift ex" -> "oc ex"
