+++ pkg/cmd/openshift/openshift.go
@@ -20,16 +20,8 @@ import (
 	"github.com/openshift/origin/pkg/cmd/server/start"
 	"github.com/openshift/origin/pkg/cmd/templates"
 	cmdutil "github.com/openshift/origin/pkg/cmd/util"
-	"github.com/openshift/origin/pkg/oc/admin"
-	diagnostics "github.com/openshift/origin/pkg/oc/admin/diagnostics"
-	sync "github.com/openshift/origin/pkg/oc/admin/groups/sync/cli"
-	"github.com/openshift/origin/pkg/oc/admin/validate"
 	"github.com/openshift/origin/pkg/oc/cli/cmd"
 	"github.com/openshift/origin/pkg/oc/cli/util/clientcmd"
-	"github.com/openshift/origin/pkg/oc/experimental/buildchain"
-	configcmd "github.com/openshift/origin/pkg/oc/experimental/config"
-	"github.com/openshift/origin/pkg/oc/experimental/dockergc"
-	exipfailover "github.com/openshift/origin/pkg/oc/experimental/ipfailover"
 )
 
 var (
@@ -103,7 +95,6 @@ func NewCommandOpenShift(name string) *cobra.Command {
 
 	startAllInOne, _ := start.NewCommandStartAllInOne(name, out, errout)
 	root.AddCommand(startAllInOne)
-	root.AddCommand(newExperimentalCommand("ex", name+" ex"))
 	root.AddCommand(newCompletionCommand("completion", name+" completion"))
 	root.AddCommand(cmd.NewCmdVersion(name, f, out, cmd.VersionOptions{PrintEtcdVersion: true, IsServer: true}))
 	root.AddCommand(cmd.NewCmdOptions(out))
@@ -112,39 +103,6 @@ func NewCommandOpenShift(name string) *cobra.Command {
 	templates.ActsAsRootCommand(root, []string{"options"})
 
 	return root
-}
-
-func newExperimentalCommand(name, fullName string) *cobra.Command {
-	out := os.Stdout
-	errout := os.Stderr
-
-	experimental := &cobra.Command{
-		Use:   name,
-		Short: "Experimental commands under active development",
-		Long:  "The commands grouped here are under development and may change without notice.",
-		Run: func(c *cobra.Command, args []string) {
-			c.SetOutput(out)
-			c.Help()
-		},
-		BashCompletionFunction: admin.BashCompletionFunc,
-	}
-
-	f := clientcmd.New(experimental.PersistentFlags())
-
-	experimental.AddCommand(validate.NewCommandValidate(validate.ValidateRecommendedName, fullName+" "+validate.ValidateRecommendedName, out, errout))
-	experimental.AddCommand(exipfailover.NewCmdIPFailoverConfig(f, fullName, "ipfailover", out, errout))
-	experimental.AddCommand(dockergc.NewCmdDockerGCConfig(f, fullName, "dockergc", out, errout))
-	experimental.AddCommand(buildchain.NewCmdBuildChain(name, fullName+" "+buildchain.BuildChainRecommendedCommandName, f, out))
-	experimental.AddCommand(configcmd.NewCmdConfig(configcmd.ConfigRecommendedName, fullName+" "+configcmd.ConfigRecommendedName, f, out, errout))
-	deprecatedDiag := diagnostics.NewCmdDiagnostics(diagnostics.DiagnosticsRecommendedName, fullName+" "+diagnostics.DiagnosticsRecommendedName, out)
-	deprecatedDiag.Deprecated = fmt.Sprintf(`use "oc adm %[1]s" to run diagnostics instead.`, diagnostics.DiagnosticsRecommendedName)
-	experimental.AddCommand(deprecatedDiag)
-	experimental.AddCommand(cmd.NewCmdOptions(out))
-
-	// these groups also live under `oc adm groups {sync,prune}` and are here only for backwards compatibility
-	experimental.AddCommand(sync.NewCmdSync("sync-groups", fullName+" "+"sync-groups", f, out))
-	experimental.AddCommand(sync.NewCmdPrune("prune-groups", fullName+" "+"prune-groups", f, out))
-	return experimental
 }
 
 var (
